feat: added new structure
