+++ .prettierrc.json
@@ -1,0 +1,6 @@
+{
+  "semi": true,
+  "trailingComma": "es5",
+  "singleQuote": true,
+  "printWidth": 80
+}
