+++ eslint.config.js
@@ -1,0 +1,32 @@
+import eslint from '@eslint/js';
+import tseslint from 'typescript-eslint';
+import prettier from 'eslint-plugin-prettier';
+
+export default tseslint.config(
+  {
+    ignores: ['dist', 'node_modules', 'eslint.config.js', 'vite.config.ts'],
+  },
+  eslint.configs.recommended,
+  ...tseslint.configs.recommended,
+  {
+    languageOptions: {
+      globals: {
+        document: 'readonly',
+        window: 'readonly',
+        Image: 'readonly',
+        requestAnimationFrame: 'readonly',
+        cancelAnimationFrame: 'readonly',
+        console: 'readonly',
+      },
+    },
+    plugins: {
+      prettier,
+    },
+    rules: {
+      'prettier/prettier': 'error',
+      'no-case-declarations': 'off',
+      'prefer-const': 'warn',
+      '@typescript-eslint/no-explicit-any': 'warn',
+    },
+  }
+);
